chore(menu-bar): reorganized menu bar layout components
