+++ src/layouts/workbench-window/application-header/menu-bar/MenuBar.tsx
@@ -1,5 +1,5 @@
-import MenuBarFileButton from "@/components/MenuBarFileButton/MenuBarFileButton";
-import MenuBarViewButton from "@/components/MenuBarViewButton/MenuBarViewButton";
+import MenuBarFileButton from "@/layouts/workbench-window/application-header/menu-bar/file-list/MenuBarFileButton";
+import MenuBarViewButton from "@/layouts/workbench-window/application-header/menu-bar/view-list/MenuBarViewButton";
 
 export default function MenuBar() {
   return (
