+++ src/layouts/workbench-window/application-header/menu-bar/view-list/MenuBarViewButton.tsx
